reference/legacy_flask/portfolio_v2/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
pandas
python-dateutil
openpyxl
